adding property and few bug fix

+++ src/CalenderEndPoints.ts
@@ -1,8 +1,14 @@
 import { runSafely } from "./Code";
 import { CalenderService } from "./service/CalenderService";
+import { UiService } from "./service/UiService";
 import { DateUtil } from "./util/DateUtil";
 
 function deleteSelectedCalenderEvents(): void {
+    if (!UiService.doesUserReConfirmedAction("This will delete the events from calender," +
+        " and can be found in calender trash.\nOnly 50 events will be deleted in one go.\n" +
+        "Proceed to continue.")) {
+        return;
+    }
     runSafely((): void => {
         new CalenderService().deleteSelectedEvent(50);
     });
@@ -16,17 +22,15 @@ function sync_todays_events(): void {
 
 function sync_currentWeek_events(): void {
     runSafely((): void => {
-        let date = new Date();
-        let weekDay = date.getDay();
-        new CalenderService().syncEvent(0 - weekDay, 6 - weekDay);
+        new CalenderService().syncEvent(DateUtil.getNumOfDaysBeforeWeekStarted(),
+            DateUtil.getNumOfDaysAfterWeekEnds());
     });
 }
 
 function sync_current_prev_next_Week_events(): void {
     runSafely((): void => {
-        let date = new Date();
-        let weekDay = date.getDay();
-        new CalenderService().syncEvent(0 - weekDay - 7, 6 - weekDay + 7);
+        new CalenderService().syncEvent(DateUtil.getNumOfDaysBeforeWeekStarted() - 7,
+            DateUtil.getNumOfDaysAfterWeekEnds() + 7);
     });
 }
 
@@ -48,6 +52,12 @@ function sync_before_30days_events(): void {
 function sync_before_90days_events(): void {
     runSafely((): void => {
         new CalenderService().syncEvent(-90, 0);
+    });
+}
+
+function scheduleInvite(): void {
+    runSafely((): void => {
+        new CalenderService();
     });
 }
